+++ src/model/DichVu.java
@@ -1,5 +1,6 @@
 package model;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 public @Data
@@ -7,5 +8,5 @@ class DichVu {
 
     String maDV;
     String tenDV;
-    double giaDV;
+    BigDecimal giaDV;
 }
