+++ src/model/HopDong.java
@@ -1,5 +1,6 @@
 package model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import lombok.Data;
 
@@ -12,7 +13,7 @@ class HopDong {
     String hoTen;
     Date ngayTao;
     String trangThai;
-    double giaThue;
+    BigDecimal giaThue;
     String Thang;
     String trangThaiThu;
 }
